requirements.txt
# Pin these dependencies
Django>=2.2,<2.3
requests==2.13.0
requests-oauthlib==0.7.0
thrift==0.10.0
thrift_connector==0.24
djangorestframework==3.10.3
django-webpack-loader==0.6.0

wagtail>=2.7,<2.8
wagtailfontawesome==1.2.1

jupyter==1.0.0
papermill==1.0.1

biopython==1.77

-e git+https://github.com/apache/airavata-django-portal-sdk.git@master#egg=airavata-django-portal-sdk
-e git+https://github.com/apache/airavata.git@develop#egg=airavata-python-sdk&subdirectory=airavata-api/airavata-client-sdks/airavata-python-sdk
-e "."
